+++ drivers/misc/mediatek/conn_soc/drv_wlan/mt_wifi/wlan/mgmt/rlm_obss.c
@@ -1,18 +1,4 @@
 /*
-* Copyright (C) 2011-2014 MediaTek Inc.
-* 
-* This program is free software: you can redistribute it and/or modify it under the terms of the 
-* GNU General Public License version 2 as published by the Free Software Foundation.
-* 
-* This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; 
-* without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
-* See the GNU General Public License for more details.
-*
-* You should have received a copy of the GNU General Public License along with this program.
-* If not, see <http://www.gnu.org/licenses/>.
-*/
-
-/*
 ** $Id: //Department/DaVinci/BRANCHES/MT6620_WIFI_DRIVER_V2_3/mgmt/rlm_obss.c#2 $
 */
 
@@ -420,8 +406,7 @@ rlmObssScanTimeout (
                 /* Skip OBSS scan */
                 prBssInfo->u2ObssScanInterval = 0;
                 
-                DBGLOG(RLM, INFO, ("WFD is running. Stop net[%u] OBSS scan.\n",
-					(UINT32)prBssInfo->ucNetTypeIndex));
+                DBGLOG(RLM, INFO, ("WFD is running. Stop net[%lu] OBSS scan.\n", prBssInfo->ucNetTypeIndex));
                 
                 return;
             }
