+++ drivers/misc/mediatek/conn_soc/drv_wlan/mt_wifi/wlan/mgmt/roaming_fsm.c
@@ -1,18 +1,4 @@
 /*
-* Copyright (C) 2011-2014 MediaTek Inc.
-* 
-* This program is free software: you can redistribute it and/or modify it under the terms of the 
-* GNU General Public License version 2 as published by the Free Software Foundation.
-* 
-* This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; 
-* without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
-* See the GNU General Public License for more details.
-*
-* You should have received a copy of the GNU General Public License along with this program.
-* If not, see <http://www.gnu.org/licenses/>.
-*/
-
-/*
 ** $Id:
 */
 
@@ -153,8 +139,8 @@ roamingFsmInit (
     P_ROAMING_INFO_T prRoamingFsmInfo;
     P_CONNECTION_SETTINGS_T prConnSettings;
 
-    DBGLOG(ROAMING, LOUD, ("->roamingFsmInit(): Current Time = %lu\n",
-		(unsigned long)kalGetTimeTick()));
+    DBGLOG(ROAMING, LOUD, ("->roamingFsmInit(): Current Time = %u\n",
+		kalGetTimeTick()));
 
     prRoamingFsmInfo = (P_ROAMING_INFO_T)&(prAdapter->rWifiVar.rRoamingInfo);
     prConnSettings = &(prAdapter->rWifiVar.rConnSettings);
@@ -183,8 +169,7 @@ roamingFsmUninit (
 {
     P_ROAMING_INFO_T prRoamingFsmInfo;
 
-    DBGLOG(ROAMING, LOUD, ("->roamingFsmUninit(): Current Time = %lu\n",
-		(unsigned long)kalGetTimeTick()));
+    DBGLOG(ROAMING, LOUD, ("->roamingFsmUninit(): Current Time = %ld\n", kalGetTimeTick()));
 
     prRoamingFsmInfo = (P_ROAMING_INFO_T)&(prAdapter->rWifiVar.rRoamingInfo);
 
@@ -212,8 +197,7 @@ roamingFsmSendCmd (
     P_ROAMING_INFO_T prRoamingFsmInfo;
     WLAN_STATUS rStatus;
 
-    DBGLOG(ROAMING, LOUD, ("->roamingFsmSendCmd(): Current Time = %lu\n",
-		(unsigned long)kalGetTimeTick()));
+    DBGLOG(ROAMING, LOUD, ("->roamingFsmSendCmd(): Current Time = %ld\n", kalGetTimeTick()));
 
     prRoamingFsmInfo = (P_ROAMING_INFO_T)&(prAdapter->rWifiVar.rRoamingInfo);
 
@@ -257,8 +241,7 @@ roamingFsmScanResultsUpdate (
     /* Check Roaming Conditions */
     ROAMING_ENABLE_CHECK(prRoamingFsmInfo);
 
-    DBGLOG(ROAMING, LOUD, ("->roamingFsmScanResultsUpdate(): Current Time = %lu\n",
-		(unsigned long)kalGetTimeTick()));
+    DBGLOG(ROAMING, LOUD, ("->roamingFsmScanResultsUpdate(): Current Time = %ld\n", kalGetTimeTick()));
 
     GET_CURRENT_SYSTIME(&prRoamingFsmInfo->rRoamingDiscoveryUpdateTime);
 
@@ -380,8 +363,7 @@ roamingFsmRunEventStart (
         return;
     }
 
-    DBGLOG(ROAMING, EVENT, ("EVENT-ROAMING START: Current Time = %lu\n",
-		(unsigned long)kalGetTimeTick()));
+    DBGLOG(ROAMING, EVENT, ("EVENT-ROAMING START: Current Time = %ld\n", kalGetTimeTick()));
 
     /* IDLE, ROAM -> DECISION */
     /* Errors as DECISION, DISCOVERY -> DECISION */
@@ -425,8 +407,7 @@ roamingFsmRunEventDiscovery (
     /* Check Roaming Conditions */
     ROAMING_ENABLE_CHECK(prRoamingFsmInfo);
 
-    DBGLOG(ROAMING, EVENT, ("EVENT-ROAMING DISCOVERY: Current Time = %lu\n",
-		(unsigned long)kalGetTimeTick()));
+    DBGLOG(ROAMING, EVENT, ("EVENT-ROAMING DISCOVERY: Current Time = %ld\n", kalGetTimeTick()));
 
     /* DECISION -> DISCOVERY */
     /* Errors as IDLE, DISCOVERY, ROAM -> DISCOVERY */
@@ -480,8 +461,7 @@ roamingFsmRunEventRoam (
     /* Check Roaming Conditions */
     ROAMING_ENABLE_CHECK(prRoamingFsmInfo);
 
-    DBGLOG(ROAMING, EVENT, ("EVENT-ROAMING ROAM: Current Time = %lu\n",
-		(unsigned long)kalGetTimeTick()));
+    DBGLOG(ROAMING, EVENT, ("EVENT-ROAMING ROAM: Current Time = %ld\n", kalGetTimeTick()));
 
     /* IDLE, ROAM -> DECISION */
     /* Errors as IDLE, DECISION, ROAM -> ROAM */
@@ -526,8 +506,8 @@ roamingFsmRunEventFail (
     /* Check Roaming Conditions */
     ROAMING_ENABLE_CHECK(prRoamingFsmInfo);
 
-    DBGLOG(ROAMING, EVENT, ("EVENT-ROAMING FAIL: reason %x Current Time = %lu\n",
-		(UINT32)u4Param, (unsigned long)kalGetTimeTick()));
+    DBGLOG(ROAMING, EVENT, ("EVENT-ROAMING FAIL: reason %x Current Time = %u\n",
+		(UINT32)u4Param, kalGetTimeTick()));
 
     /* IDLE, ROAM -> DECISION */
     /* Errors as IDLE, DECISION, DISCOVERY -> DECISION */
@@ -571,8 +551,7 @@ roamingFsmRunEventAbort (
 
     ROAMING_ENABLE_CHECK(prRoamingFsmInfo);
 
-    DBGLOG(ROAMING, EVENT, ("EVENT-ROAMING ABORT: Current Time = %lu\n",
-		(unsigned long)kalGetTimeTick()));
+    DBGLOG(ROAMING, EVENT, ("EVENT-ROAMING ABORT: Current Time = %ld\n", kalGetTimeTick()));
 
     eNextState = ROAMING_STATE_IDLE;
     /* IDLE, DECISION, DISCOVERY, ROAM -> IDLE */
@@ -603,8 +582,7 @@ roamingFsmProcessEvent (
     IN P_ROAMING_PARAM_T prParam
     )
 {
-    DBGLOG(ROAMING, LOUD, ("ROAMING Process Events: Current Time = %lu\n",
-		(unsigned long)kalGetTimeTick()));
+    DBGLOG(ROAMING, LOUD, ("ROAMING Process Events: Current Time = %ld\n", kalGetTimeTick()));
 
     if (ROAMING_EVENT_DISCOVERY == prParam->u2Event) {
         roamingFsmRunEventDiscovery(prAdapter, prParam->u2Data);
