+++ drivers/misc/mediatek/conn_soc/drv_wlan/mt_wifi/wlan/mgmt/wnm.c
@@ -1,18 +1,4 @@
 /*
-* Copyright (C) 2011-2014 MediaTek Inc.
-* 
-* This program is free software: you can redistribute it and/or modify it under the terms of the 
-* GNU General Public License version 2 as published by the Free Software Foundation.
-* 
-* This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; 
-* without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
-* See the GNU General Public License for more details.
-*
-* You should have received a copy of the GNU General Public License along with this program.
-* If not, see <http://www.gnu.org/licenses/>.
-*/
-
-/*
 ** $Id: //Department/DaVinci/TRUNK/MT6620_5931_WiFi_Driver/mgmt/wnm.c#1 $
 */
 
@@ -193,8 +179,7 @@ wnmRunEventTimgingMeasTxDone (
     ASSERT(prAdapter);
     ASSERT(prMsduInfo);
 
-    DBGLOG(WNM, LOUD, ("EVENT-TX DONE: Current Time = %lu\n",
-		(unsigned long)kalGetTimeTick()));
+    DBGLOG(WNM, LOUD, ("EVENT-TX DONE: Current Time = %ld\n", kalGetTimeTick()));
 
     prStaRec = cnmGetStaRecByIndex(prAdapter, prMsduInfo->ucStaRecIndex);
