+++ drivers/misc/mediatek/conn_soc/drv_wlan/mt_wifi/wlan/os/linux/gl_p2p_kal.c
@@ -1,18 +1,4 @@
 /*
-* Copyright (C) 2011-2014 MediaTek Inc.
-* 
-* This program is free software: you can redistribute it and/or modify it under the terms of the 
-* GNU General Public License version 2 as published by the Free Software Foundation.
-* 
-* This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; 
-* without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
-* See the GNU General Public License for more details.
-*
-* You should have received a copy of the GNU General Public License along with this program.
-* If not, see <http://www.gnu.org/licenses/>.
-*/
-
-/*
 ** $Id: @(#) gl_p2p_cfg80211.c@@
 */
 
@@ -1248,7 +1234,7 @@ kalP2PIndicateMgmtTxStatus (
                 (pucFrameBuf == NULL) ||
                 (u4FrameLen == 0)) {
             DBGLOG(P2P, TRACE, ("Unexpected pointer PARAM. 0x%lx, 0x%lx, %u.",
-				(UINT_32)prGlueInfo, (UINT_32)pucFrameBuf, (UINT32)u4FrameLen));
+				(UINT_32)prGlueInfo, (UINT_32)pucFrameBuf, u4FrameLen));
             ASSERT(FALSE);
             break;
         }
